refactor(unstyled): move value utilities to un/value
Move the type and map helper out of unstyled-utilities into a
new unstyled-value module and update imports across packages.

- Replace imports of Value and mapValue from "./unstyled-utilities" with
  "./unstyled-value" inside packages/unstyled (cva & tests).
- Update package exports in packages/unstyled/package.json to expose
  "./value": "./src/unstyled-value.ts" instead of the old utilities path.
- Adjust cross-package import in packages/design to reference
  "unstyled/value" instead of "unstyled/utilities".

This clarifies module responsibilities by separating generic value helpers
from other unstyled utilities and prepares the codebase for clearer
exports and future reorganization. Tests and consumers are updated to
reflect the new export paths.
